Refactor seeder files to use testing
